+++ src/components/DashBoard/DashBoard.jsx
@@ -1,5 +1,5 @@
 import { Navigation } from '../Navigation/Navigation';
-import { Outlet } from 'react-router';
+
 import { Balance } from 'components/Balance/Balance';
 import { Carrency } from 'components/Carrency/Carrency';
